requirements.txt
flask
python-dotenv
openperplex
openai
google-generativeai
